refactor: react query cache key 상수 분리
