+++ frontend/src/components/CategoryModifyModal/CategoryModifyModal.tsx
@@ -12,7 +12,7 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY } from '@/constants';
+import { CACHE_KEY } from '@/constants/api';
 
 import categoryApi from '@/api/category';
