+++ frontend/src/components/FilterCategoryItem/FilterCategoryItem.tsx
@@ -9,7 +9,7 @@ import { userState } from '@/recoil/atoms';
 
 import Button from '@/components/@common/Button/Button';
 
-import { CACHE_KEY } from '@/constants';
+import { CACHE_KEY } from '@/constants/api';
 import { CATEGORY_TYPE } from '@/constants/category';
 import { PALETTE } from '@/constants/style';
