+++ frontend/src/components/Profile/Profile.tsx
@@ -11,7 +11,8 @@ import { userState } from '@/recoil/atoms';
 import Button from '@/components/@common/Button/Button';
 import Fieldset from '@/components/@common/Fieldset/Fieldset';
 
-import { CACHE_KEY, PATH } from '@/constants';
+import { PATH } from '@/constants';
+import { CACHE_KEY } from '@/constants/api';
 import { CONFIRM_MESSAGE } from '@/constants/message';
 
 import { createPostBody } from '@/utils';
