+++ frontend/src/constants/api.ts
@@ -5,10 +5,24 @@ const API = {
 
 const API_URL = process.env.API_URL;
 
+const CACHE_KEY = {
+  AUTH: 'auth',
+  CATEGORIES: 'categories',
+  CATEGORY: 'category',
+  ENTER: 'enter',
+  GOOGLE_CALENDAR: 'googleCalendar',
+  MY_CATEGORIES: 'myCategories',
+  PROFILE: 'profile',
+  SCHEDULE: 'schedule',
+  SCHEDULES: 'schedules',
+  SUBSCRIPTIONS: 'subscriptions',
+  VALIDATE: 'validate',
+};
+
 const RESPONSE = {
   STATUS: {
     UNAUTHORIZED: 401,
   },
 };
 
-export { API, API_URL, RESPONSE };
+export { API, API_URL, CACHE_KEY, RESPONSE };
