+++ frontend/src/pages/CalendarPage/CalendarPage.tsx
@@ -24,7 +24,8 @@ import ScheduleAddModal from '@/components/ScheduleAddModal/ScheduleAddModal';
 import ScheduleModal from '@/components/ScheduleModal/ScheduleModal';
 import ScheduleModifyModal from '@/components/ScheduleModifyModal/ScheduleModifyModal';
 
-import { CACHE_KEY, CALENDAR } from '@/constants';
+import { CALENDAR } from '@/constants';
+import { CACHE_KEY } from '@/constants/api';
 import { DAYS } from '@/constants/date';
 import { TRANSPARENT } from '@/constants/style';
